Address PR review feedback: race conditions, held status, bigint
- Wrap execute! in with_lock to prevent double-execution race condition
- Add "held" as a terminal status for hold actions so they don't get
  picked up by .due on every sweep
- Move RefundUnpaidPurchasesWorker.perform_async outside the lock/transaction
  so it only fires after the DB commit succeeds
- Guard refund action with user.suspended? check to prevent silent no-ops
- Change payout_amount_cents to bigint to avoid overflow on large balances
- Remove unused JS imports (request, assertResponseError, ResponseError,
  showAlert)

+++ db/migrate/20261119011939_create_scheduled_payouts.rb
@@ -10,7 +10,7 @@ class CreateScheduledPayouts < ActiveRecord::Migration[7.1]
       t.string :status, null: false, default: "pending"
       t.bigint :created_by_id
       t.datetime :executed_at
-      t.integer :payout_amount_cents
+      t.bigint :payout_amount_cents
       t.timestamps
 
       t.index :status
